requirements.txt
django
djangorestframework
jsonfield
pyyaml
jsonschema
drf-spectacular
